// src/types/pages/productCard/ProductCard_T.ts
export type ProductCard_T = {
    productName:            string;
    description:            string;
    newProduct:             boolean;
    productUrl:            string;
    categoryImages: {
        mobileUrl:          string;
        tabletUrl:          string;
        desktopUrl:         string;
    }
}